fix(asr): silence CodeQL float-overflow alert in clusterer test
CodeQL flagged tests/asr_speaker_clusterer_test.cpp:26 (high): `x * x` computes
in float and may overflow before promotion to the double accumulator. Cast the
first operand to double so the multiply happens in double precision. The
production norm/dot paths that accumulate in double already do this
(SpeakerEmbedder, SpeakerClusterer::centroid norm, AsrSegmenter); the float
accumulators (dot/energy) aren't affected — no widening conversion, so no alert.

+++ tests/asr_speaker_clusterer_test.cpp
@@ -23,7 +23,7 @@ std::vector<float> vec(int axis, float perturb)
     v[axis] = 1.0f;
     v[(axis + 3) % 8] = perturb;
     double n = 0;
-    for (float x : v) n += x * x;
+    for (float x : v) n += static_cast<double>(x) * x; // accumulate in double
     n = std::sqrt(n);
     for (float& x : v) x = static_cast<float>(x / n);
     return v;
